Python script for 3 -> 2 colour converison
This is a python script in beta phase. Please make a backup of the E-Paper-Master folder before executing.
It's intended for users who want to try the software on the 2-colour E-Paper display from waveshare. Download and copy this script in the 'Calendar' folder. Follow the instructions within the python script and edit accordingly.

After executing the script, reboot the Raspberry Pi to apply the changes. That's all. Any feedback is greatly appreciated.