More user- and E-Paper friendly
Some parts have been improved in this version (1st Oct 2018). The main one is that the E-Paper display is now put to sleep. Waveshare did not mention for a long time that not doing so will eventually damage the screen. The sleep mode has been implemented now.

Apart from this, the main details can now be inserted at the top of the file and don't have to be searched up anymore. Some debug messages have also been added, though they are uncommented. That's all.