Cleaned up the code & added explainations
This update is nothing major. It splits the main programm in 2 files, one for the programm itself (now containing mostly functions) and a new file, icon_positions_locations, containing all icons and their respective locations (path) as well as the position on the E-Paper Display).
Apart from that, new notes have been added which help to understand what the script is trying to achieve.