Added bugfix for iCalendar and truncating events
Contains some improvements in the section tha filters events from the iCalendar URL. The sorting algorithm has been improved to display events after today and in chronological order. If the event names is too long for the line, it'll be truncated until it fits.
Special thanks and credit to Hubert for suggesting the above mentioned improvements.
Also fixed an issue where the iCalendar would throw errors if the Alarm action for an event was set to 'None'.