Added experimental RSS feed parser
This is an RSS parser, written in python. It's in development phase so it might not work as expected. Once it does, the aim is to display either events, To-Do's or RSS feeds in the section below the Calendar template. Testers are most welcome. If you find any bugs/improvement ideas, feel free to write me a mail. Thanks in advance