Added images for calibration
In the calibration module, previously an image with only one colour was first generated converted and then displayed. To decrease the calibration time a little, the updated module does not generate the images, but rather opens them from the 'other' path. As the image does not need to be generated each time, calibration now works a little faster.