Optimising Installer for new dependencies
Previously, the installer would only install dependencies when choosing the install option. The installer will install dependencies when updating as well. The installer will now ask before attempting to update the operating system when choosing the updating/installing option. When installing/updating, only those dependencies will be installed, which are missing on the system. This check is done by test-importing the module in python3. If it's not found, it's installed.