version1.5 -> 1.6 with lots of improvements
The installer now has version v1.6 and contains some optimisations during installation and updating. When selecting the update or install option in the installer, it will ask whether or not you want to update the system. Dependencies will be installed when selecting the update or install option. This helps reduce the updating time, fixes problems with missing dependencies and makes the installation process a bit faster. Also added the option to 'update-anyway' when selecting the update option which bypasses the check for the configuration file in the Calendar folder. Apart from that, some important information will be displayed with a small delay on the Terminal instead of all at once. Lastly, colours have been added on some output lines. This version of the installer has been tested on a Raspberry Pi Zero W running on Raspbian Lite form April 2019.