Lots of improvements
E-Paper is now split into three sections: top, main and bottom. Users can now choose what to display at which section. With this change, the main script will only execute the code required to generate the user-specified design. Better text handling. Text is automatically aligned in the middle of the specified box (unless specified otherwise) using the default font (unless specified otherwise). Added Agenda-View as an option in the main section, which will show upcoming events and times in the next few days. It's length is adaptive. If the bottom section is empty, it will reach to the bottom, otherwise it will reach to the bottom section. RSS-feeds will not show the titles, but rather a short summary (for more detailed information). Removed some import warnings as they are handled by the new installer (v1.6). Also made the code a bit cleaner and added a few more descriptions above some code sections.