Stick to ics v.0.4 for now
A new release (v0.5) is out for ics (the library parsing the iCalendar events) but as with any new release, there are still bugs and fixes that need to be handled first. For now, the installer will stick to v0.4. When all tests have been passed, an upgrade to ics v0.5 will be considered.