Please don't use this yet
This is the first commit for release v1.7. It was adjusted to work on the new release (v1.7). The code has been cleaned up and runs faster now. It's also possible to run the software even with a different username than 'pi'. Supervisor handles launching at boot more efficiently and gives up running the main file if the software won't run for at least 30 seconds.