Minor improvements + Bugfixes
Fixed a bug where the update interval would be calculated incorrectly. The display will now update as specified in the settings file.
Instead of importing a single driver file, this module will automatically import the driver file for the specified E-Paper (in the settings file). This means that even though more E-Papers are now supported, the code will not become too complex.

Calibration has been omitted for now to test if the E-Paper shows signs of ghosting.