added more conditions all-day events must meet
I further defined the conditions an event must meet in order to be categorized as "all-day". The old test was causing our script to categorize some events that start at 0000UTC (2000EST, for me) as all-day. This didn't happen to all events, but I dove into the ics files and noticed a couple of cases where this would happen.

Now it must meet the following criteria:
1. Start at 00:00 UTC
2. End at 00:00 UTC
3. The end date must be greater than or equal to the start date plus one day

I feel pretty confident that the above three criteria will be satisfactory, but I will continue testing and find more bugs.