fixed bug that would edit timezone when not needed
Unfortunately ical doesn't give you any info about whether an event is an all-day event and it also doesn't include timezone info, which is different than all other events. It's important to know which events are all-day and which are not, because all-day events are not converted to local time, instead the local time tag is only added so it shows up in the proper order. Previously, when an event started at 0000 UTC, inkycal would assume it was an all-day event. So, other events that started at 0000 UTC (which, for me, is 7 pm EST), would be handled in the same way as all-day events and they would not be converted to local time. So that meant all my 1900 EST events would show up as starting at 0000 EST. I added code that attempts to further clarify what an all-day event is and ensure that regular events starting at 0000 UTC are properly converted to local time.

Unfortunately, the code is not nearly as clean and beautiful as the code that ace wrote. I attempted to write is as clear as possible, so that future people will be able to understand it. This update is critical for those in timezones who frequently have events that start at 0000 UTC.