added a scaling factor for the weather icons
As ace and I discussed a while back (mid-2020?) the unicode weather icons report an incorrect size in the y direction. Some of the icon's sizes are off by enough that the human eye can see that they are no longer centered in the y direction. During another project of mine, I went through each icon and added a manual scaling factor that, when used along with the font size, can help us properly center the icons. I had difficulty integrating that scaling factor into the weather module because it's a very complicated program and ace wrote it. But, I believe that by editing the y alignment line in the draw_icon fuction to factor in this size correction, all the icons are now correctly centered. That appears to be the case with my 12" screen, although I haven't yet tested it with the 7.5" screen.