Fix for deleted groups causing Exceptions
Deleted groups would previously cause the module to fail as the api would still show todos from groups that have been deleted, but not the name of the group.
Iterating over the groups caused a KeyError Exception. This commit does an additional check to see if the group really does exist or has been deleted.