main: add image hashing
build md5 sum over the resulting assembled image(s) and
check against a stored hash file to determine if we really need
to update the screen.
Option can be controlled by new image_hash global option.

If info_section is enabled while image_hash is on,
the time of update will be stripped from the section,
as it is clearly not hashable.

In the end this enables us to update the information in the
background way more frequent without increasing the stress
caused to the display

Signed-off-by: Konrad Weihmann <kweihmann@outlook.com>